proto-grpc: home the status bounding shared by runtime crates
`MAX_STATUS_MESSAGE_LEN` and its helpers live in `runtime-next`, but three
other crates map errors onto the same wire and need the same ceiling. Move
them to `proto-grpc`, which each of those crates already depends on, and
re-export from `runtime-next` so callers are unchanged.

No behavior change: `runtime-next` bounds exactly what it bounded before.
